--- api/get_test.go
@@ -1,36 +1,0 @@
-package api
-
-import (
-	"io/ioutil"
-	"testing"
-
-	"github.com/google/go-cmp/cmp"
-	"github.com/qri-io/dataset"
-)
-
-func TestGetZip(t *testing.T) {
-	run := NewAPITestRunner(t)
-	defer run.Delete()
-
-	// Save a version of the dataset
-	ds := run.BuildDataset("test_ds")
-	ds.Meta = &dataset.Meta{Title: "some title"}
-	ds.Readme = &dataset.Readme{ScriptBytes: []byte("# hi\n\nthis is a readme")}
-	run.SaveDataset(ds, "testdata/cities/data.csv")
-
-	// Get a zip file binary over the API
-	dsHandler := NewDatasetHandlers(run.Inst, false)
-	gotStatusCode, gotBodyString := APICall("/get/peer/test_ds?format=zip", dsHandler.GetHandler)
-	if gotStatusCode != 200 {
-		t.Fatalf("expected status code 200, got %d", gotStatusCode)
-	}
-
-	// Compare the API response to the expected zip file
-	expectBytes, err := ioutil.ReadFile("testdata/cities/exported.zip")
-	if err != nil {
-		t.Fatalf("error reading expected bytes: %s", err)
-	}
-	if diff := cmp.Diff(string(expectBytes), gotBodyString); diff != "" {
-		t.Errorf("byte mismatch (-want +got):\n%s", diff)
-	}
-}
