+++ api/render.go
@@ -23,11 +23,6 @@ func NewRenderHandlers(inst *lib.Instance) *RenderHandlers {
 
 // RenderHandler renders a given dataset ref
 func (h *RenderHandlers) RenderHandler(w http.ResponseWriter, r *http.Request) {
-	if r.Method == "OPTIONS" {
-		apiutil.EmptyOkHandler(w, r)
-		return
-	}
-
 	p := &lib.RenderParams{
 		Ref:       HTTPPathToQriPath(r.URL.Path[len("/render"):]),
 		OutFormat: "html",
