+++ api/search_test.go
@@ -21,7 +21,6 @@ func TestSearchHandlers(t *testing.T) {
 	}
 
 	searchCases := []handlerTestCase{
-		{"OPTIONS", "/", nil},
 		// TODO (b5): lol wut Get requests don't have bodies
 		{"GET", "/", mustFile(t, "testdata/searchRequest.json")},
 		{"DELETE", "/", nil},
